new branch for api changes.
split the Documents table into two:
we still retain a Documents table, but now Documents have multiple
components which have an associated image that we will perform OCR on and
then index.

changing up the routes accordingly:
there will be a main /upload route that will handle uploading
a document with a single component, presumably for a well-formed image
on which OCR will work well.

there is also a /upload/components route that can handle uploading
multiple images (components of a document). these multiple images
are presumably sections of a main image, where each section is amenable
to OCR.